[addons][vfs] prevent C++ designed Kodi Header IFileTypes.h
On this commit becomes them replaced to use independent enum and
structures, to pass related data between Kodi and Addon.

This prevents conflicts when using "C" and ensures a secure ABI interface.

In addition, it is much safer in the event that the Kodi changes what in its
headers. This can easily occur because previously it was not directly visible
that IFileTypes.h was also used on addons.

Only currently still used on libXBMC_addon.h and removed on PVR rework.