Proposal to increase queue slightly
This proposal is based on a change already implemented by @davilla for
MrMC
which also fixes known issues with specific HLS streams.

Whereas @davilla's changes for MrMC go a lot further than we need to fix
the known issues increasing the SetMaxTimeSize of Audio *and* Video
demux queues from a value of 8 seconds to 10 seconds is sufficient.
@peak3d proposed 12 seconds for both.

While these streams have issues, we are in no position to fix those
streams. Other tools behave perfectly fine on these streams as well and
with this small fix the streams behave well in Kodi as well.

For Matrix we are looking into a more extensive fix in the player, but
while we do not expect large changes in the VideoPlayer in Leia to be
accepted and we like to see this fixed in the next release of Leia we
are proposing this here.