[addons][audioengine] change interface code to match "C"
This changes the code where work between Kodi and addon to correct
"C" where also allow (in future) to use in "C" only.

There on few functions the "struct" added where "C" needed, further
is a separate class added where handle AudioEngineFormat itself and
them before is renamed to AUDIO_ENGINE_FORMAT.
This was needed to remove the C++ contructor and functions inside structure
where not match to "C".