[addons][audioengine] separate Kodi and Addon enum parts
This add own enums on addon interface to identify channels and stream types.
That becomes be by 4 new functions translated between Kodi and Addon.

The enums become futher on next changes a detailed documentation to have
better on addons.

Also are this thought to use in future on AudioDecoder interface and if
comes (maybe) back on audio dsp.

Most mandatory background of this is to make 100% sure, that if on Kodi's
AudioEngine is something changed, that the addon not become a mistmatch
with them.