Application: Async player start/stop
In order to fix a deadlock when users
start/stop very fast, we offload player
close into a background job as part of a
jobqueue.
In order to make sure there are no races
between subsequent opens, we also put it
in the same jobqueue.
That makes sure the playerstop is really
done before the next open can happen.

Meanwhile the app stays responsive to the
user, because nothing blocks the mainthread.