VideoPlayer: Uninitialize RenderManager in OnExit
As CloseFile can now be called from background
threads, we run into deadlocks between the
background worker and VP, because both want
stuff to be done on the mainthread at cleanup.

Moving the render uninit into the OnExit method
makes sure only the VP thread does cleanup that
needs to be done on the mainthread.