[ios] Runtime fix after spdlog merge
IOS init process is not considered a standard Kodi init process.
This Log call is now being called prior to the Logger creation in CApplication::Create()

This is more a quick fix to get running builds. Further investigation of how DarwinEmbedded
platforms init would need to be done to potentially fix/change the init process to insure
CApplication::Create() is called earlier in the process, and windowing init is done in
windowing instead of outside prior as is the current case.