[addons][pvr] change stream properties and do over inputstream
Before was them in inputstream and the same as copy in PVR, this now remove it
there to use also the defines from inputstream.
This to make in case of change more sure to have equal and to have it cleaner.

Further is the documentation a bit changes to match the rework part.