[peripherals] Re-introduce "input.enablejoystick" setting
This setting was removed in the joystick subsystem introduced in Kodi 17
Krypton. Instead, controller input could be disabled from the GUI by
disabling the peripheral.joystick add-on.

The reason for re-adding this setting is due to Kodi's interaction with the
window manager. On systems without a window manager, Kodi won't receive an
event when the main app loses focus. By introducing this setting, an API is
provided to clients so they can appropriately control controller input based
on external knowledge.