[addons][docs] extend C++ related initial webpage
There becomes a table about usable types added, as this is currently
the primary way of binary addons should also them listed on first place.

This becomes extended in future but for now better as before, where
not direct visible for what is.

The classes before was only in:
- "Language Development" -> "C++" -> "Interface" - "kodi::addon" -> "Addon type instances"
where was to long.

Now in:
- "Language Development" -> "C++"