[addons][global] add setting read way for enum's and create docs
By them a new way to get and set a setting via enum is added. Some
addons use enum's and before was it not so nice how it was handled.
Now with the template it can define a wanted type.

Further is with them the related documentation about the settings
functions added.