Lower repo update frequency to 4 hours
In a VOD-centric world, Kodi add-ons are living by the grace of the VOD
content provider and the stability of the VOD infrastructure.
Unfortunately, since Kodi add-ons are not first-class citizens, they are
not prepared or made aware of coming changes to the VOD infrastructure.

This means that "suddenly" the Kodi add-on stops working when the VOD
infrastructure undergoes changes that the Kodi add-on was not prepared
for. **This immediately impacts users.** And no matter how fast add-on
developers react, or how well they are improving their processes, this
will always be reactive.

Many add-ons pro-actively test their add-ons at full length and are made
aware of such issues as soon as they arise, and a fix is out as fast as
possible. We still need to wait until the add-on is being accepted and
merged in the Kodi repository, And additionally it takes up to 24 hours
to get these add-ons installed on end-user systems.

24 hours mean that daily users are always affected, which is a real
problem for Kodi as a reliable platform.

- Mean time to detection (MTTD) is now close to 0
- Mean time to acknowledge (MTTA) depends on add-on development team, could be within hours
- Mean time to recovery (MTTR) is a problem
  - Fixing depends on add-on development team
  - Getting it merged into Kodi repository depends on Team Kodi
  - Getting it on user systems depends on this 24h check in Kodi

We would like to see this 24 hour check in Kodi to be reduced to 4 or 6
hours so that fixes can be delivered before they impact end-users.

In most cases being able to get the fix on end-users systems by 18:00
would avoid impacting the majority of users.