[addons][vfs] prevent direct use of "void*" on addon interface
There becomes the "void*" usage replaced to a "C define VFS_FILE_HANDLE" and "C++ using VFSFileHandle".
Also is a from addon used memory pointer changed from void* to uint8_t*.
On other parts where only relate to Kodi is changed to KODI_HANDLE.

So is on addon not direct a void* used and more clean on other languages.