[addons] Allow repo server to dictate update interval
We currently have no possibility to update the repository update
interval; it is fixed at 24 h. It is desirable to control this value
from the server side so we are more flexible and can, e.g., offer faster
updates if server load allows it.

This commit adds client-side support for the `X-Kodi-Recheck-After`
response header to the checksum HTTP request. It specifies how many
seconds to wait until the repo is checked for updates the next time.
The server is free to set this value as it sees fit. If the header is
not included in the response, a default value of 24 h is used.

The header is only supported in repositories that have checksums enabled
for all directories.