[addons][audiodecoder] improve error handling about CAudioDecoder::Init
Before was addon after his init call not checked.
In case it has returned false was  still Kodi's side of data set. But
this then not always good, as we not know what it brings.

Further was "int channels" and "int sampleRate" not set before where breaks
e.g. coverity scan.

This change does this:
- Make translate of addon to Kodi only if addon returned true
- Set "int channels" and "int sampleRate" to -1
- In case addon returned true and wanted values are not set, becomes a error reported and Kodi's function returns false