[addons][cmake][binary] add test about "C" ABI
This introduces cmake-based test system to the addon headers.
The background to this is to have the header contained therein in the correct "C" ABI.
This is not yet called up by Kodi's creation itself, as there are still many errors in it.

The background is easy to check whether it is correct and if everything has been reworked to be carried out during Kodi's creation to avoid new mistakes.

Why this cmake is in this folder has another background, regarding trademark policy history it could come that the dev-kit folder is independent and not directly in Kodi.
In addition, you have a better overview of what the new "abi-interface-test.cmake" belongs to and is not mixed with Kodi's other cmake stuff.

Why the header must be so explicit in "C" has the following reasons:
- As a basis for other languages
- With more security if Kodi and Addon were created by different compilers and versions