[addons][binary] add "C" include guards to his headers
On the "C" headers inside dev kit was before only the "#pragma once" used.
Them is nice and from most compilers supported, but it's not a standard.

This adds also the old and everywhere supported guard macros to this headers. Thought
for cases that this code is taken for other languages, where have a own
compiler about and not support the wanted "pragma".

See description here https://en.wikipedia.org/wiki/Pragma_once