[filesystem][music] fix music tag set if file contains several streams
This solves problems with files that contain multiple streams.

Occurs primarily with some audio decoder add-ons and especially with SACD streams (which can be used for popular music and contains large tag data which must be given to Kodi).

Before the "tag" of the basic source (e.g. SACD ISO itself) was used and all things contained therein were given, so the necessary names of the respective content were missing.