[addons][database] remove no more used "broken" list
There was stil a table inside where was last used in Kodi 16 Jarvis and
replaced on Kodi 17 Krypton with add of "metadata" way.

Was coming to them by implement another type for "deprecated" and looked
about change for them there and seen not needed.

This commit remove his unused code and update addon database to 29 with
remove of this.