[addons][vfs] fix addon version update
Before was by a VFS addon update still the old version shown and the new
visible after Kodi restart.

This was coming with the binary addon instance handling and here where
the new was created before the old was destroyed, in this case does
the system create a new instance on already opened addon (which in this
case the old version).

This change unload now before the VFS addon (if not used somewhere else)
and create the new after them, with this the correct CAddonDll with new
version becomes created.

Further is the way improved and only the changed addon restarted, all
others stays untouched. Before was by update of one addon all VFS
addons restarted.