[addons][install] fix origin view direct after install of addon
Before was if a new addon was installed and shown via "User installed add-ons"
the origin (Repository) not known and shown as "Unknown", after a Kodi restart
it was then available.

There was before after the install by CAddonInstallJob the manager called to add
the new, but after the new was added there, the installer set the origin
in database where was before not available in manager.
With this the database was correct and on next Kodi start used.

This change move the database update to the manager so that it also direct update
the origin within his listed addons.