[addons] replace addon.xml <broken> with <lifecyclestate ...>
This is used to take advantage of several different types of status.
With `<broken>` before, only a yes or no (if this is not available) could be given.

The counterpart to `<broken>Some Text</broken>` would now be:
- `<lifecyclestate type="broken">Some Text</lifecyclestate>`

In addition, this now also allows:
- `<lifecyclestate type="deprecated">Some Text</lifecyclestate>`
  - To note addon complete usable, but deprecated and maybe replaced by another one
- `<lifecyclestate type="normal">Some Text</lifecyclestate>`
  - To show as normal, this not really needed as an addon without always "normal"
  - Maybe good to use, if the normal addon has a special condition where good to show in GUI

Another change to the previous one is that specified texts can now also be entered in different languages using e.g. `lang="en_GB"`.