[addons][peripheral] use C++ classes for addon function values
This changes the addon functions that they used direct the C ++ classes for their values.
It is intended to keep "C" away from "C++" addons as much as possible and to have the interface more secure.

Because these classes were already available, it was easy to convert and the respective work on the addon is no longer necessary.