[addons][gui] add dialog Yes/No for ask use of deprecated/broken addon
This adds a display of a Yes/No dialog which is shown when the user wants to activate an outdated or broken addon using the GUI (context menu / addon dialog).

He is then asked whether it should really be activated and is shown add-on text which describes why it is out of date or broken.

This function is added in a new file / class because it is used in several places (currently 2)
and may use other places in the future, and similar things can be added to the new class to keep a better overview.