[addons][gui] make kodi<->addon interface header "C" conform
On the one hand, this moves the "C" API header of the GUI to the new "kodi/c-api/gui" folder.

Furthermore it has been revised to have "C" compatible and every `void*` part now has a matching `typedef` to make it easier to identify the associated position.
As the only typedef with the "KODI_GUI_CLIENT_HANDLE" also a `using ClientHandle = KODI_GUI_CLIENT_HANDLE;` added to have it in the corresponding addon function.

Als einzigtes typedef with zum "KODI_GUI_CLIENT_HANDLE" auch ein `using ClientHandle = KODI_GUI_CLIENT_HANDLE;`
In the following commits, the C ++ headers are also revised in order not to be in conflict if included in "C".
The associated documentation will also be revised in the following commits.