[addons][gui] fix/improve window control creation
Before was the control not available until the `OnInit()` by addon was called.
This has produced a ugly work way on addons as the control has need always a "new"
call within this `OnInit()` and the delete of control must be somewhere outside.

Also was by this not possible to use them as fixed mem inside his window class.

By this change becomes the window init called on first control creation to have
his related skin xml loaded and available for wanted control creation.