[addons][platform] change documentation to have more usable
Before was it not good viewed inside doxygen generated HTML as the
functions was viewed outside of class name.

This change to define class name and increase a bit platform specific group descriptions.

Also was wrongly before a empty `@param[in]` used where brings conflict to coming `@return`

On the API itself is by this and commit before nothing in code work changed and not needs new versions.