[vfs][addons][audiodecoder] continue addon search if open of extension failed
This thought for the case that a addon tried with a extension where can also be on other types.

Before has it returned the function and no further search to find process part.

The error and fix specially visible by *.iso's and use of audiodecoder.sacd with vfs.libarchive (where iso support added).

Before has it called the audiodecoder where fails on a normal file ISO and returned the scan function as unsupported.
With this change here is the search continued and try then also with other addons and with Kodi's own VFS support.