XBDateTime: Use signed LARGE_INTEGER instead of unsigned ULARGE_INTEGER
It is logically incorrect to use ULARGE_INTEGER in CDateTime and
CDateTimeSpan, because the QuadPart can be negative (for example,
assigning the negative timzeone offset in seconds to a CDateTimeSpan).

We also rename private member functions ToULargeInt() and FromULargeInt()
to ToLargeInt() and FromLargeInt().

Changes were made using following shell commands:

   sed -i 's/ULARGE_INTEGER/LARGE_INTEGER/g' xbmc/XBDateTime.*
   sed -i 's/ToULargeInt(/ToLargeInt(/g' xbmc/XBDateTime.*
   sed -i 's/FromULargeInt(/FromLargeInt(/g' xbmc/XBDateTime.*