[addons][tools] add own StringUtils.h to help addons about
This is the part most often used by p8-platform to edit strings in C++.
With this, it is added to the DevKit with.

It's based 99% on that of Kodi and is more or less a copy of it.
Unfortunately Kodi's own cannot be used in addons because it then breaks the "C" ABI.

It works completely independently of Kodi, so there are a few fewer functions in it than in the original (but these were not on p8-platform either).

The background is the confusion that currently exists with p8-platform. This is based on Kodi,
but with its last release (October 26th, 2016) it is very old and brings problems with some OS.
This in turn resulted in a clone from Kodi which contains fixes.
Now p8-platform will be continued again (without a release) which in turn does not correspond
to our necessary changes and does not make direct sense on the external side (e.g. TVOS).
Linux distributions continue to use the release of it and we need an extension to it for all OS.

Therefore it is brought into Kodi's addon header so that it is always up to date.

In addition, other format functions (e.g. `kodi::Log()`) in the dev kit can then also be used by the new system.