[addons][tools] add helper header for thread processing
This adds another help header to easily create threads on addons and is based entirely
on the thread support that has existed since C++11, therefore OS independent.

It is a further step to remove the p8-platform usage of addons. As a result of the fact
that Kodi and addons C++17 is now coming and p8-platform is based on C++, correct use
cannot be guaranteed, as it is not certain whether there will also be C++17.

Code is also partly taken from Kodi.