GUIComponent: Remove register of stereoscopic manager
The stereoscopic manager is registered in 2 places.
Once in GUICompontent's init and another time in load
skin where all other AddMsgTarget calls take place.

During startup of the application following sequence happens:
* GuiCompontent init
* Skin unload
* Skin load

When we unload the skin all message targets are removed and
they are readded in skin load.
Hence the registration in GUIComponent is irrelevant.