[addons] add on 3 "C" related headers the ifdef's for header include
On three headers was still the ifdef's about alternative of "pragma once"
missing. This add them.

Is good to have as "pragma once" not a standardized part and possible that
not supported everywhere, the "ifdef" instead works always.

This only be on "C" related headers as them maybe base for other languages.