TestDateTime: fix GetAsStringsWithBias and add Tzdata tests
The 'GetAsStringsWithBias' test logic assumed the timezone on
machine bullding Kodi is set to 'UTC'. This is not always the case,
plus hour value was declared with a typo.

To address this, we compare outputs of two runs of date library,
and instead add a 'Tzdata' test case comparing local date and time
against computed values from different timezones:

  find /usr/share/zoneinfo/Etc/ -type f | sort -Vu | sed 's,^/usr/share/zoneinfo/,,' | while read TZ
  do
    TMSTR=$(LANG=C TZ="$TZ" date '+%Y-%m-%dT%H:%M:%S%Ez' -d "1991-05-14 12:34:56 UTC" | sed 's/[0-9][0-9]$/:&/')
    echo "  // LANG=C TZ=\"$TZ\" date '+%Y-%m-%dT%H:%M:%S%Ez' -d \"1991-05-14 12:34:56 UTC\" | sed 's/[0-9][0-9]$/:&/'"
    echo "  tps = date::floor<std::chrono::seconds>(dateTime.GetAsTimePoint());"
    echo "  zone = date::make_zoned(\"$TZ\", tps);"
    echo "  EXPECT_EQ(date::format(\"%FT%T%Ez\", zone), \"$TMSTR\") << \"tzdata information not valid for '$TZ'\";"
    echo ""
  done

Signed-off-by: Vasyl Gello <vasek.gello@gmail.com>