[addons][gui] show compile warning if GLonDX.h is used without Win/Angle
This add a warning message to report use on wrong OS with it.

This header need always Windows OS and a Angle installed on depends.
One goal with it is also to use on cmake test script to check correctness of all
addon headers, as this becomes then also included on e.g. Linux makes no sense to use
"#error ..." for it.