[OSX][packaging] allow the ability to create dmg from xcode generated project
currently the mkdmg-osx.sh script will fail when project is an xcode generated project

  You need to specify the build target

The issue is that xcode projects arent a set configuration type at cmake generation time
CORE_BUILD_CONFIG will be empty. This passes the xcode configuration type to the script
so the type can be used to build the dmg correctly.