[build] fix CMake-related build quoting issue
CMake on my Debian 10 machine was quoting the PATH environment variable
being set here, which made the shell interpret that not as a variable
being set but as a command, which it then would print a confusing error
message about not being able to find.

We work around this by using CMake's built-in environment modification
utility.