Incorrect ALIASES specified
The ALIASES has 2 items where the double quotes are not properly escaped and thus the specific ALIAS ends ant the first non escaped double quote.
In 1.8.17 no warning is issued though whether the results are OK is doubtful.
In 1.9.1 a warning is given.

Fix: Properly escaped double quotes inside ALIASES.