[win32] Fixed CMake out-of-source build on Windows
CMake is smart enough to detect symlink on Windows, but at the same time
is not smart enough to copy symlink itself or linked content on Windows.
Currently it fails to copy 'freebsd.xml' which is symlink to 'linux.xml'.
Suggested workaround uses CMake internal command and avoids copy of
unchanged file.