windowing/gbm/drm: default to original crtc mode
This brings RES_DESKTOP handling of gbm in line with other windowing
systems.

If the framebuffer has been set up with a valid mode before kodi
startup then use that as the default / DESKTOP mode, otherwise fall
back to the previous behaviour of setting DESKTOP to the preferred
mode reported by drm.

This avoids unnecessary mode switches at kodi startup (which usually
lead to screen blanking for a short time) and allows users to easily
override broken video/edid/... setups by manually configuring a mode
in linux which then kodi uses too. That is especially important when
running kodi for the first time, if it switches to a different mode
that might not be working with the user's monitor the result is a
black screen with no easy possibility to fix the settings.

Signed-off-by: Matthias Reichl <hias@horus.com>