[rendering] Fix OverlayRenderer alpha blending
If `glBlendFunc` is used, it sets the blend mode for RGB as well as
the alpha channel. Using GL_SRC_ALPHA and GL_ONE_MINUS_SRC_ALPHA for
the alpha channel leads to the equation

Alpha out = Overlay alpha * Overlay alpha +
            (1 - Overlay alpha) * Destination alpha

This sort of blending does not make much sense for the alpha channel and
means that the alpha of the framebuffer can become < 1 even if it was
completely opaque (destination alpha = 1 everywhere) before.

On Wayland, we promise the compositor that our surface is opaque. If
alpha is not 1.0 at some pixel, display artifacts may occur, so the
incorrect alpha channel leads to incorrect display.

To fix this problem, the alpha equation is changed to:

Alpha out = Overlay alpha + (1 - Overlay alpha) * Destination alpha

This removes the squaring of the overlay alpha and guarantees that
alpha out == 1 if destination alpha == 1.