[vfs][addons] Don't check for subdirectories in an open audio decoder list
Previously "CFileDirectoryFactory::Create" tried to find and open additional paths with an open audio decoder file, which contains sub-songs.
The sub-files always contain a ".***stream" as an extension, with this the file extension is checked for "stream".

However, since other folders can never be contained there, this check is avoided.