[addons] Regarding addon.xml "<news>", use changelog.txt as an alternative
In the event that the changes log (news) in addon.xml is empty, check whether it is an installed addon and read a changelog.txt as a replacement, if available.

Many addons use this "changelog.txt", but less use the "<news>" field in addon.xml, and nobody actually uses the translation path in it (once unknown and difficult to keep correct in the case of continuous changes).
Hence, this brings a fall back to reading the text file as a replacement.