[addons] Add used class headers directly
The background to this is to insert some add-on related headers directly in the respective file and to avoid problems when removing or changing at external locations (which previously contained these headers). This turns a small thing into a larger request.

With this commit, future work will be reduced a little.