[addons][audiodecoder] improve audio channel list handover
This improves the value for transferring the channel list supported by the add-on at the init call of the audio decoder.

Once you save the memory in Kodi's class and furthermore I had problems managing the "const enum AudioEngineChannel ** info" correctly when setting up "C" and its tests, because it refers to a fixed array, works with the new value now.