[addons][audiodecoder] always use size_t when reading data from files to array.
This is a mandatory guideline on the revised API!

It must never be used as such `uint8_t* buffer, int size`, since the subsequent value size cannot be assigned to buffer with "int", so an auto-creation records this as a single value referring to the external location. In order to recognize this correctly, it must always be set with a size_t after its value (never before).