[addons][audiodecoder] use typedef's for "C" API function pointers
The background to this is that this value is required in several places in the Sandbox API and is easier to manage using typedef.

It also makes the structure clearer to identify available functions.

I left the _V1 in there for the time being, because it would comb anyway and to leave out subsequent changes.
These commits and requests are mainly about having the finale cleaner and only recognizing the changes that are directly related to it.