[addons][audiodecoder] pass "void*" value as addon target handle
The background is that in a sandbox system it is no longer possible to simply exchange a structure defined as a pointer between Kodi and addon, since Kodi never edits and uses this, it should always be passed as "void*" for something like that.

In addition, it becomes smaller because the target handle / class pointer add-on is then transferred directly and no longer the current API base structure.