[addons][audiodecoder] replace string use on AUDIO_DECODER_INFO_TAG and rename struct
In this commit, the structure AUDIO_DECODER_INFO_TAG is renamed once to KODI_ADDON_AUDIODECODER_INFO_TAG. The background is that for me a "_" is preferred for group division, and can also be used as a division of paths. So that a "C" structure roughly corresponds to a C ++ namespace.

Furthermore, the more primary change in it, whereby the fixed array size is removed from strings and re-created using "strdup".

This then has the following advantages:
- The string length is irrelevant and does not need to be taken into account by e.g. lyrics
- Storage space is saved because unnecessary areas of the fixed memory do not have to be set
- It allows easier handling in upcoming sandbox basis systems