Properly link with -pthread instead of -lpthread
This patch changes the way cmake handles
Thread support to correctly link with -pthread instead of -lpthread.
That way GCC automatically pulls in -latomic if needed. I have noticed
that debian/rules already forces LDFLAGS to contain -latomic, however
it appears too early in the linker command to be early (it needs to be
after the objects needing this library).