[addons][audiodecoder] rename extension part "stream" to "_adecstrm"
This is an important change to avoid possible conflicts with other file extensions.
The name "stream" was previously too general and may cause problems if something else uses the same thing.

In addition, its name is now as a define in addon header in order to be able to find used places in Kodi more easily and to keep it smaller when changes are made.