[addons][audiodecoder] make static API function names on C++ header equal to C
This change has to keep the function names of the "C" part the same as the static am on the C++, only "ADDON_" remains as the beginning, because I suspected that with e.g. "init" there could be problems in other places.

Another reason is the planned future automation of the interface between Kodi and addon, where it will be automatically added to the C ++ header. Hence, names must be kept the same.