[addons] make SSIZE_T available for all OS via addon_bash.h
This adds the necessary macro in addon_base.h.

The reason for this is that the revised API will have stricter guidelines and, for example, a size value of a memory must always appear as `size_t` or` ssize_t`, otherwise an auto-creation script cannot recognize in which cont

It is currently also available at filesystem.h, but will continue to work due to checks and will be removed there with API rework.