[gui] Allow tabs to text in GUI
This change introduces the support of tabs in texts, which are then correctly displayed in the GUI with up to 4 spaces.

The new text variable for this is `[TABS] tab amount [/TABS]`, or direct use of the `\t`.

The background is to have larger text blocks more structured, as individual spaces are more difficult to lay in lines in the correct vertical order.