[skin][estuary] allow scroll within select dialog
This allows scrolling text on the selection dialog, this is to support larger text in it.

To make it easier to see, the sizes had to be increased slightly, otherwise it would recognize the second line outside of its text block and always scroll when two lines were used.

The background to this is the addition of available support for file extensions starting with audiodecoder addon.