[addons] fix extension update in case about addon change
Previously "insert" was used there alone. If changes are made, this does not cause any function, as the given identification is already set! Therefore an addon update / installation / deinstallation could never work in this context.

Here it is changed to use [] as set.

This means that a VFS or audio decoder add-on can be used immediately after its installation and no longer needs Kodi restart.