[addons][audiodecoder] Add help function to read out the desired track number
This adds a new static function in the C ++ header of the audio decoder addon.

This is intended to read the track number desired by Kodi from the called path. This is added here because actually all addons that support tracks use exactly the same code, so now in the header to avoid there.