[addons][audiodecoder] rename audio_decoder.h to audiodecoder.h
With the newly planned sandbox / API level system it becomes necessary to have the "C" files matching its namespace.

As an example, a "C" function in audiodecoder must use "kodi_addons_audiodecoder_some_function ()" as the name, with which the "_" acts as a namespace divider to begin with, if now, as before, it would see "audio" and "decoder" as separate areas.

Advantageous for the automatic creation script, otherwise problems can arise for him.