[addons][cmake] Clean up addon header CMakeLists.txt
Previously, "core_add_library" was used which, once referring to library by name, and on the other hand always had the path as "addons_kodi-dev-kit_include_ *" in it.

This adds a new CMake function with which it is done via "core_add_devkit_header". The difference is that this checks on "ENABLE_STATIC_LIBS" and always adds "addons_kodi-dev-kit_include_" as begin.

The following background is behind it:
- So that the kodi-dev-kit path is not set in its parts (here CMakeLists.txt) and the folder can simply be moved to another location.
- The "core_add_devkit_header" could possibly be replaced with another function if kodi-dev-kit were used independently of Kodi.
- Better to use with auto scripts

The other `if (HEADERS)` has the background if inline cmake calls were integrated into it and HEADERS would be given as empty.